requirement.txt
streamlit
scikit-learn
pandas
plotly
requests
python-dotenv
google-genai


# requirements.txt

streamlit
pandas
scikit-learn
python-dotenv
google-generativeai